requirements.txt
autopep8==1.5.3
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
coverage==5.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
Flask-Testing==0.8.0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.13.0
six==1.15.0
SQLAlchemy==1.3.17
toml==0.10.1
Werkzeug==1.0.1
